Refactor pure functions out of TaskTable.jsx and into TaskTableColumns.jsx
