+++ client/app/queue/utils.js
@@ -572,3 +572,9 @@ export const cityForRegionalOfficeCode = (code) => {
 
   return regionalOffice ? regionalOffice.city : COPY.UNKNOWN_REGIONAL_OFFICE;
 };
+
+export const hasDASRecord = (task, requireDasRecord) => {
+  return (task.appeal.isLegacyAppeal && requireDasRecord) ? Boolean(task.taskId) : true;
+};
+
+export const collapseColumn = (requireDasRecord) => (task) => hasDASRecord(task, requireDasRecord) ? 1 : 0;
